refactor(weather): move the Open-Meteo archive client into pkg/weather
The historical-conditions lookup was private to pkg/chatbot, reachable only
through the !weather command. The onscreens location feed needs the same answer
to fill the rotators' $weather variable, so the client moves to its own
stdlib-only package and chatbot keeps just the injection seam.
